# public/docs/api-docs.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Laravel",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "https://api-soporte-ti.onrender.com/api"
        }
    ],
    "paths": {
        "/compras": {
            "get": {
                "operationId": "compras.index",
                "tags": [
                    "Compra"
                ],
                "responses": {
                    "500": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string",
                                            "const": "Error al obtener las compras."
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "success",
                                        "message",
                                        "error"
                                    ]
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Compra"
                                            }
                                        }
                                    },
                                    "required": [
                                        "success",
                                        "data"
                                    ]
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "operationId": "compras.store",
                "tags": [
                    "Compra"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CompraRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "500": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string",
                                            "const": "Error al registrar la compra."
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "success",
                                        "message",
                                        "error"
                                    ]
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string",
                                            "const": "Compra registrada correctamente."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Compra"
                                        }
                                    },
                                    "required": [
                                        "success",
                                        "message",
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Compra": {
                "type": "object",
                "properties": {
                    "id_compra": {
                        "type": "integer"
                    },
                    "n_documento": {
                        "type": "string"
                    },
                    "id_proveedor": {
                        "type": "integer"
                    },
                    "id_usuario": {
                        "type": "integer"
                    },
                    "total_compra": {
                        "type": "string"
                    },
                    "forma_pago": {
                        "type": "string"
                    },
                    "observacion": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "fecha": {
                        "type": "string"
                    },
                    "estado": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "date-time"
                    }
                },
                "required": [
                    "id_compra",
                    "n_documento",
                    "id_proveedor",
                    "id_usuario",
                    "total_compra",
                    "forma_pago",
                    "observacion",
                    "fecha",
                    "estado",
                    "created_at",
                    "updated_at"
                ],
                "title": "Compra"
            },
            "CompraRequest": {
                "type": "object",
                "properties": {
                    "n_documento": {
                        "type": "string",
                        "maxLength": 50
                    },
                    "id_proveedor": {
                        "type": "integer"
                    },
                    "id_usuario": {
                        "type": "integer"
                    },
                    "total_compra": {
                        "type": "number",
                        "minimum": 0
                    },
                    "forma_pago": {
                        "type": "string",
                        "maxLength": 50
                    },
                    "observacion": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "fecha": {
                        "type": "string",
                        "format": "date"
                    },
                    "estado": {
                        "type": [
                            "boolean",
                            "null"
                        ]
                    },
                    "detalles": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "id_repuesto": {
                                    "type": "integer"
                                },
                                "cantidad": {
                                    "type": "integer",
                                    "minimum": 1
                                },
                                "precio": {
                                    "type": "number",
                                    "minimum": 0
                                },
                                "sub_total": {
                                    "type": "number",
                                    "minimum": 0
                                },
                                "estado": {
                                    "type": "boolean"
                                },
                                "id_detalle_compra": {
                                    "type": "integer"
                                }
                            },
                            "required": [
                                "id_repuesto",
                                "cantidad",
                                "precio",
                                "sub_total"
                            ]
                        },
                        "minItems": 1
                    }
                },
                "required": [
                    "n_documento",
                    "id_proveedor",
                    "id_usuario",
                    "total_compra",
                    "forma_pago",
                    "fecha",
                    "detalles"
                ],
                "title": "CompraRequest"
            },
...
        },
        "responses": {
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            }
        }
    }
}